New binary protocol implementation
This merges new server & binary protocol implementations.  The new
protocol is more efficient and allows to properly encode all EdgeDB data
types.  The Python driver implementation is in a separate repo:

    https://github.com/edgedb/edgedb-python

This is still alpha-quality.  The immediate next steps are as follows:

* Make EdgeDB repl use the binary protocol;

* Ensure that con.fetch() and con.execute() work correctly with
  transactions and DDL;

* Implement support for savepoints;

* Implement a process pool for compilers to speedup connection time;

* Drop old server and client implementations.

The following changes in this commit are worth mentioning:

* `api/errors.txt` defines all EdgeDB errors.  The `edb gen-errors` script
  is used to generate Python classes for server and client
  implementations;

* Most of user-facing exceptions were changed to errors from
  `api/errors.txt`;

* New 'edb gen-types' script generates Python code for well-known
  EdgeDB types;

* Named and positional-only query parameters are now fully supported.

+++ setup.py
@@ -24,9 +24,11 @@ import shutil
 import subprocess
 import textwrap
 
+from distutils import extension as distutils_extension
 from distutils.command import build as distutils_build
+from distutils.command import build_ext as distutils_build_ext
 
-from setuptools import setup
+import setuptools
 from setuptools.command import develop as setuptools_develop
 
 
@@ -40,6 +42,13 @@ RUNTIME_DEPS = [
     'pygments',
     'setproctitle',
     'typing_inspect~=0.3.1',
+    'uvloop~=0.12.0rc1',
+]
+
+CYTHON_DEPENDENCY = 'Cython==0.29.0'
+
+BUILD_DEPS = [
+    CYTHON_DEPENDENCY,
 ]
 
 EXTRA_DEPS = {
@@ -54,6 +63,15 @@ EXTRA_DEPS = {
         'requests-xml',
     ],
 }
+
+EXT_CFLAGS = ['-O2']
+EXT_LDFLAGS = []
+
+
+if platform.uname().system != 'Windows':
+    EXT_CFLAGS.extend([
+        '-fsigned-char', '-Wall', '-Wsign-compare', '-Wconversion'
+    ])
 
 
 def _compile_parsers(build_lib, inplace=False):
@@ -103,7 +121,9 @@ def _compile_build_meta(build_lib, pg_config):
         f.write(content)
 
 
-def _compile_postgres(build_base):
+def _compile_postgres(build_base, *,
+                      force_build=False, fresh_build=True,
+                      run_configure=True, build_contrib=True):
 
     proc = subprocess.run(
         ['git', 'submodule', 'status', 'postgres'],
@@ -131,7 +151,9 @@ def _compile_postgres(build_base):
     else:
         build_stamp = None
 
-    if source_stamp != build_stamp:
+    is_outdated = source_stamp != build_stamp
+
+    if is_outdated or force_build:
         system = platform.system()
         if system == 'Darwin':
             uuidlib = 'e2fs'
@@ -140,27 +162,39 @@ def _compile_postgres(build_base):
         else:
             raise NotImplementedError('unsupported system: {}'.format(system))
 
-        if postgres_build.exists():
+        if fresh_build and postgres_build.exists():
             shutil.rmtree(postgres_build)
         build_dir = postgres_build / 'build'
-        build_dir.mkdir(parents=True)
-        subprocess.run([
-            str(postgres_src / 'configure'),
-            '--prefix=' + str(postgres_build / 'install'),
-            '--with-uuid=' + uuidlib,
-        ], check=True, cwd=str(build_dir))
+        if not build_dir.exists():
+            build_dir.mkdir(parents=True)
+
+        if run_configure or fresh_build or is_outdated:
+            subprocess.run([
+                str(postgres_src / 'configure'),
+                '--prefix=' + str(postgres_build / 'install'),
+                '--with-uuid=' + uuidlib,
+            ], check=True, cwd=str(build_dir))
+
         subprocess.run(
             ['make', '-j', str(max(os.cpu_count() - 1, 1))],
             cwd=str(build_dir), check=True)
-        subprocess.run(
-            ['make', '-C', 'contrib', '-j', str(max(os.cpu_count() - 1, 1))],
-            cwd=str(build_dir), check=True)
+
+        if build_contrib or fresh_build or is_outdated:
+            subprocess.run(
+                [
+                    'make', '-C', 'contrib', '-j',
+                    str(max(os.cpu_count() - 1, 1))
+                ],
+                cwd=str(build_dir), check=True)
+
         subprocess.run(
             ['make', 'install'],
             cwd=str(build_dir), check=True)
-        subprocess.run(
-            ['make', '-C', 'contrib', 'install'],
-            cwd=str(build_dir), check=True)
+
+        if build_contrib or fresh_build or is_outdated:
+            subprocess.run(
+                ['make', '-C', 'contrib', 'install'],
+                cwd=str(build_dir), check=True)
 
         with open(postgres_build_stamp, 'w') as f:
             f.write(source_stamp)
@@ -246,10 +280,118 @@ class develop(setuptools_develop.develop):
         super().run(*args, **kwargs)
 
 
-setup(
+class build_postgres(setuptools.Command):
+
+    description = "build postgres"
+
+    user_options = [
+        ('configure', None, 'run ./configure'),
+        ('build-contrib', None, 'build contrib'),
+        ('fresh-build', None, 'rebuild from scratch'),
+    ]
+
+    def initialize_options(self):
+        self.configure = False
+        self.build_contrib = False
+        self.fresh_build = False
+
+    def finalize_options(self):
+        pass
+
+    def run(self, *args, **kwargs):
+        _compile_postgres(
+            pathlib.Path('build').resolve(),
+            force_build=True,
+            fresh_build=self.fresh_build,
+            run_configure=self.configure,
+            build_contrib=self.build_contrib)
+
+
+class build_ext(distutils_build_ext.build_ext):
+
+    user_options = distutils_build_ext.build_ext.user_options + [
+        ('cython-annotate', None,
+            'Produce a colorized HTML version of the Cython source.'),
+        ('cython-directives=', None,
+            'Cython compiler directives'),
+    ]
+
+    def initialize_options(self):
+        # initialize_options() may be called multiple times on the
+        # same command object, so make sure not to override previously
+        # set options.
+        if getattr(self, '_initialized', False):
+            return
+
+        super(build_ext, self).initialize_options()
+
+        if os.environ.get('EDGEDB_DEBUG'):
+            self.cython_always = True
+            self.cython_annotate = True
+            self.cython_directives = "linetrace=True"
+            self.define = 'PG_DEBUG,CYTHON_TRACE,CYTHON_TRACE_NOGIL'
+            self.debug = True
+        else:
+            self.cython_always = False
+            self.cython_annotate = None
+            self.cython_directives = None
+
+    def finalize_options(self):
+        # finalize_options() may be called multiple times on the
+        # same command object, so make sure not to override previously
+        # set options.
+        if getattr(self, '_initialized', False):
+            return
+
+        import pkg_resources
+
+        # Double check Cython presence in case setup_requires
+        # didn't go into effect (most likely because someone
+        # imported Cython before setup_requires injected the
+        # correct egg into sys.path.
+        try:
+            import Cython
+        except ImportError:
+            raise RuntimeError(
+                'please install {} to compile edgedb from source'.format(
+                    CYTHON_DEPENDENCY))
+
+        cython_dep = pkg_resources.Requirement.parse(CYTHON_DEPENDENCY)
+        if Cython.__version__ not in cython_dep:
+            raise RuntimeError(
+                'edgedb requires {}, got Cython=={}'.format(
+                    CYTHON_DEPENDENCY, Cython.__version__
+                ))
+
+        from Cython.Build import cythonize
+
+        directives = {
+            'language_level': '3'
+        }
+
+        if self.cython_directives:
+            for directive in self.cython_directives.split(','):
+                k, _, v = directive.partition('=')
+                if v.lower() == 'false':
+                    v = False
+                if v.lower() == 'true':
+                    v = True
+
+                directives[k] = v
+
+        self.distribution.ext_modules[:] = cythonize(
+            self.distribution.ext_modules,
+            compiler_directives=directives,
+            annotate=self.cython_annotate,
+            include_path=["edb/server2/pgproto/"])
+
+        super(build_ext, self).finalize_options()
+
+
+setuptools.setup(
     setup_requires=[
         'setuptools_scm',
-    ] + RUNTIME_DEPS,
+    ] + RUNTIME_DEPS + BUILD_DEPS,
     use_scm_version=True,
     name='edgedb-server',
     description='EdgeDB Server',
@@ -259,7 +401,9 @@ setup(
     include_package_data=True,
     cmdclass={
         'build': build,
+        'build_ext': build_ext,
         'develop': develop,
+        'build_postgres': build_postgres,
     },
     entry_points={
         'console_scripts': [
@@ -267,6 +411,25 @@ setup(
             'edgedb-server = edb.server.main:main',
         ]
     },
+    ext_modules=[
+        distutils_extension.Extension(
+            "edb.server2.pgproto.pgproto",
+            ["edb/server2/pgproto/pgproto.pyx"],
+            extra_compile_args=EXT_CFLAGS,
+            extra_link_args=EXT_LDFLAGS),
+
+        distutils_extension.Extension(
+            "edb.server2.edgecon.edgecon",
+            ["edb/server2/edgecon/edgecon.pyx"],
+            extra_compile_args=EXT_CFLAGS,
+            extra_link_args=EXT_LDFLAGS),
+
+        distutils_extension.Extension(
+            "edb.server2.pgcon.pgcon",
+            ["edb/server2/pgcon/pgcon.pyx"],
+            extra_compile_args=EXT_CFLAGS,
+            extra_link_args=EXT_LDFLAGS),
+    ],
     install_requires=RUNTIME_DEPS,
     extras_require=EXTRA_DEPS,
     test_suite='tests.suite',
